added yr_o target coordinates to the structure Entry_t, modified the termination condition from smallest length to saturatio

+++ global_consts.py
@@ -13,6 +13,6 @@ BKZ_SIEVING_CROSSOVER = 55
 N_SIEVE_THREADS = 5
 N_MAX_SLICER_ITERATIONS = 400
 NRAND_FACTOR = 5.
-DTYPE = np.float128 #np.longdouble or np.float64
+DTYPE = np.float64 #np.longdouble or np.float64
 EPS2 = 1.0001 #(1+epsilon) squared
 NPROJ_TESTS = 2048
